+++ src/API.mli
@@ -1145,6 +1145,11 @@ module Utils : sig
     ?name:string -> ?graft:([`After | `Before] * string) ->
     depth:int -> Ast.Loc.t -> Data.term -> Ast.program
 
+  (** Hackish API. Move a term from one runtime (after execution) to another
+      one which has all the needed symbols *)
+  val relocate_closed_term :
+    State.t * Data.term -> State.t -> (Data.term,string) Stdlib.Result.t
+
   (** Lifting/restriction/beta (LOW LEVEL, don't use) *)
   val move : from:int -> to_:int -> Data.term -> Data.term
   val beta : depth:int -> Data.term -> Data.term list -> Data.term